+++ src/main/kotlin/graphql/dataTransferObjects/Account.kt
@@ -1,3 +1,3 @@
 package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
-class Account(override val id: Int) : AccountData
+class Account(override val id: Int) : AccountData, ReadAccountResult