+++ src/main/kotlin/graphql/dataTransferObjects/InvalidUserId.kt
@@ -2,6 +2,6 @@ package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
 import com.neelkamath.omniChatBackend.graphql.routing.Placeholder
 
-object InvalidUserId : ReadOnlineStatusResult, CreatePrivateChatResult {
+object InvalidUserId : ReadOnlineStatusResult, CreatePrivateChatResult, ReadAccountResult {
     fun getPlaceholder() = Placeholder
 }