+++ docs/CHANGELOG.md
@@ -11,9 +11,11 @@ The entire project (i.e., the GraphQL API, REST API, and server) uses the same v
 ### Added
 
 - `Query.readAllowedEmailAddressDomains`
+- `union ReadAccountResult`
 
 ### Changed
 
+- Return a `type ReadAccountResult` instead of a `type Account` from `Query.readAccount`.
 - Remove the `lastOnline` field from `type OnlineStatus`.
 - Remove the `state` and `statuses` fields from the following:
     - `interface Message`